icn_runtime::context

Struct RuntimeContext

pub struct RuntimeContext {

    pub current_identity: Did,
    pub mana_ledger: SimpleManaLedger,
    pub pending_mesh_jobs_tx: Sender<ActualMeshJob>,
    pub pending_mesh_jobs_rx: Mutex<Receiver<ActualMeshJob>>,
    pub job_states: Arc<DashMap<JobId, JobState>>,
    pub governance_module: Arc<Mutex<GovernanceModule>>,
    pub mesh_network_service: Arc<dyn MeshNetworkService>,
    pub signer: Arc<dyn Signer>,
    pub did_resolver: Arc<dyn DidResolver>,
    pub dag_store: Arc<Mutex<dyn DagStorageService<DagBlock> + Send>>,
    pub reputation_store: Arc<dyn ReputationStore>,
    pub policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>,
    pub time_provider: Arc<dyn TimeProvider>,
    pub default_receipt_wait_ms: u64,

}

Fields
current_identity: Didmana_ledger: SimpleManaLedgerpending_mesh_jobs_tx: Sender<ActualMeshJob>pending_mesh_jobs_rx: Mutex<Receiver<ActualMeshJob>>job_states: Arc<DashMap<JobId, JobState>>governance_module: Arc<Mutex<GovernanceModule>>mesh_network_service: Arc<dyn MeshNetworkService>signer: Arc<dyn Signer>did_resolver: Arc<dyn DidResolver>dag_store: Arc<Mutex<dyn DagStorageService<DagBlock> + Send>>reputation_store: Arc<dyn ReputationStore>policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>time_provider: Arc<dyn TimeProvider>

Source of time for timestamp generation.

default_receipt_wait_ms: u64

Default timeout in milliseconds when waiting for job execution receipts.

Implementations

impl RuntimeContext

pub fn new_with_ledger_path( current_identity: Did, mesh_network_service: Arc<dyn MeshNetworkService>, signer: Arc<dyn Signer>, did_resolver: Arc<dyn DidResolver>, dag_store: Arc<TokioMutex<dyn DagStorageService<DagBlock> + Send>>, mana_ledger_path: PathBuf, reputation_store_path: PathBuf, policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>, ) -> Arc<Self>

Create a new context using a mana ledger stored at mana_ledger_path.

pub fn new_with_ledger_path_and_time( current_identity: Did, mesh_network_service: Arc<dyn MeshNetworkService>, signer: Arc<dyn Signer>, did_resolver: Arc<dyn DidResolver>, dag_store: Arc<TokioMutex<dyn DagStorageService<DagBlock> + Send>>, mana_ledger_path: PathBuf, reputation_store_path: PathBuf, policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>, time_provider: Arc<dyn TimeProvider>, ) -> Arc<Self>

Create a new context using a mana ledger stored at mana_ledger_path and the provided [TimeProvider].

pub fn new_with_mana_ledger( current_identity: Did, mesh_network_service: Arc<dyn MeshNetworkService>, signer: Arc<dyn Signer>, did_resolver: Arc<dyn DidResolver>, dag_store: Arc<TokioMutex<dyn DagStorageService<DagBlock> + Send>>, mana_ledger: SimpleManaLedger, reputation_store_path: PathBuf, policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>, ) -> Arc<Self>

Create a new context with a preconstructed mana ledger.

pub fn new_with_mana_ledger_and_time( current_identity: Did, mesh_network_service: Arc<dyn MeshNetworkService>, signer: Arc<dyn Signer>, did_resolver: Arc<dyn DidResolver>, dag_store: Arc<TokioMutex<dyn DagStorageService<DagBlock> + Send>>, mana_ledger: SimpleManaLedger, _reputation_store_path: PathBuf, policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>, time_provider: Arc<dyn TimeProvider>, ) -> Arc<Self>

Create a new context with a preconstructed mana ledger and custom [TimeProvider].

pub fn new( current_identity: Did, mesh_network_service: Arc<dyn MeshNetworkService>, signer: Arc<dyn Signer>, did_resolver: Arc<dyn DidResolver>, dag_store: Arc<TokioMutex<dyn DagStorageService<DagBlock> + Send>>, ) -> Arc<Self>

Create a new context using the default mana ledger path (./mana_ledger.sled).

pub fn new_with_paths( current_identity: Did, mesh_network_service: Arc<dyn MeshNetworkService>, signer: Arc<dyn Signer>, did_resolver: Arc<dyn DidResolver>, dag_path: PathBuf, mana_ledger_path: PathBuf, reputation_store_path: PathBuf, policy_enforcer: Option<Arc<dyn ScopedPolicyEnforcer>>, ) -> Result<Arc<Self>, CommonError>

Create a new context using filesystem paths for the DAG store and mana ledger. The store type is selected based on enabled persistence features.

pub fn new_with_stubs( current_identity_str: &str, ) -> Result<Arc<Self>, CommonError>

pub fn new_with_stubs_and_mana( current_identity_str: &str, initial_mana: u64, ) -> Result<Arc<Self>, CommonError>

pub async fn internal_queue_mesh_job( self: &Arc<Self>, job: ActualMeshJob, ) -> Result<(), HostAbiError>

pub async fn spawn_mesh_job_manager(self: Arc<Self>)

pub async fn spawn_mana_regenerator( self: Arc<Self>, amount: u64, interval: StdDuration, )

Spawn a background task that periodically credits all mana accounts.

amount is the mana added per interval. interval specifies how often the crediting occurs.

pub async fn integrity_check_once(&self) -> Result<(), CommonError>

Verify integrity of all DAG blocks once.

pub async fn spawn_integrity_checker(self: Arc<Self>, interval: StdDuration)

Spawn a background task that periodically checks DAG integrity.

pub async fn get_mana(&self, account: &Did) -> Result<u64, HostAbiError>

pub async fn spend_mana( &self, account: &Did, amount: u64, ) -> Result<(), HostAbiError>

pub async fn credit_mana( &self, account: &Did, amount: u64, ) -> Result<(), HostAbiError>

pub async fn anchor_receipt( &self, receipt: &IdentityExecutionReceipt, ) -> Result<Cid, HostAbiError>

Anchors an execution receipt to the DAG store and returns the content identifier (CID). This method is called by the Host ABI function host_anchor_receipt.

pub async fn create_governance_proposal( &self, payload: CreateProposalPayload, ) -> Result<String, HostAbiError>

pub async fn cast_governance_vote( &self, payload: CastVotePayload, ) -> Result<(), HostAbiError>

pub async fn close_governance_proposal_voting( &self, proposal_id_str: &str, ) -> Result<String, HostAbiError>

pub async fn execute_governance_proposal( &self, proposal_id_str: &str, ) -> Result<(), HostAbiError>

pub async fn delegate_vote( &self, from_did: &str, to_did: &str, ) -> Result<(), HostAbiError>

pub async fn revoke_delegation( &self, from_did: &str, ) -> Result<(), HostAbiError>

pub async fn ingest_external_proposal( &self, proposal_bytes: &[u8], ) -> Result<(), HostAbiError>

Inserts a proposal received from the network into the local GovernanceModule.

pub async fn ingest_external_vote( &self, vote_bytes: &[u8], ) -> Result<(), HostAbiError>

Inserts a vote received from the network into the local GovernanceModule.

pub async fn host_anchor_receipt( &self, receipt: &IdentityExecutionReceipt, ) -> Result<(), HostAbiError>