icn_runtime::context

Trait MeshNetworkService

pub trait MeshNetworkService:
    Send
    + Sync
    + Debug
    + DowncastSync {
    // Required methods
    fn announce_job<'life0, 'life1, 'async_trait>(
        &'life0 self,
        job: &'life1 ActualMeshJob,
    ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn announce_proposal<'life0, 'async_trait>(
        &'life0 self,
        proposal_bytes: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn announce_vote<'life0, 'async_trait>(
        &'life0 self,
        vote_bytes: Vec<u8>,
    ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn collect_bids_for_job<'life0, 'life1, 'async_trait>(
        &'life0 self,
        job_id: &'life1 JobId,
        duration: StdDuration,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<MeshJobBid>, HostAbiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn notify_executor_of_assignment<'life0, 'life1, 'async_trait>(
        &'life0 self,
        notice: &'life1 JobAssignmentNotice,
    ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn try_receive_receipt<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        job_id: &'life1 JobId,
        expected_executor: &'life2 Did,
        timeout: StdDuration,
    ) -> Pin<Box<dyn Future<Output = Result<Option<IdentityExecutionReceipt>, HostAbiError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn as_any(&self) -> &dyn Any;
}

Trait for a service that can broadcast and receive mesh-specific messages. This is the local definition for icn-runtime.

Required Methods

fn announce_job<'life0, 'life1, 'async_trait>( &'life0 self, job: &'life1 ActualMeshJob, ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn announce_proposal<'life0, 'async_trait>( &'life0 self, proposal_bytes: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn announce_vote<'life0, 'async_trait>( &'life0 self, vote_bytes: Vec<u8>, ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn collect_bids_for_job<'life0, 'life1, 'async_trait>( &'life0 self, job_id: &'life1 JobId, duration: StdDuration, ) -> Pin<Box<dyn Future<Output = Result<Vec<MeshJobBid>, HostAbiError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn notify_executor_of_assignment<'life0, 'life1, 'async_trait>( &'life0 self, notice: &'life1 JobAssignmentNotice, ) -> Pin<Box<dyn Future<Output = Result<(), HostAbiError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn try_receive_receipt<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, job_id: &'life1 JobId, expected_executor: &'life2 Did, timeout: StdDuration, ) -> Pin<Box<dyn Future<Output = Result<Option<IdentityExecutionReceipt>, HostAbiError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

fn as_any(&self) -> &dyn Any

Implementations

impl dyn MeshNetworkService

pub fn is<__T: MeshNetworkService>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: MeshNetworkService>( self: Box<Self>, ) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: MeshNetworkService>( self: Rc<Self>, ) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: MeshNetworkService>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: MeshNetworkService>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_arc<__T: MeshNetworkService + Any + Send + Sync>( self: Arc<Self>, ) -> Result<Arc<__T>, Arc<Self>>

Returns an Arc-ed object from an Arc-ed trait object if the underlying object is of type __T. Returns the original Arc-ed trait if it isn’t.