icn_runtime::error_recovery

Trait ErrorClassifier

pub trait ErrorClassifier<E> {
    // Required method
    fn is_recoverable(&self, error: &E) -> bool;

    // Provided method
    fn is_permanent(&self, error: &E) -> bool { ... }
}

Trait for classifying errors as recoverable or not

Required Methods

fn is_recoverable(&self, error: &E) -> bool

Returns true if the error is recoverable and operation should be retried

Provided Methods

fn is_permanent(&self, error: &E) -> bool

Returns true if the error indicates a permanent failure